move funding outside of bench
