+++ bench-tps/src/bench.rs
@@ -1,17 +1,13 @@
-use rayon::iter::split;
-
 use {
     crate::{
         bench_tps_client::*,
         cli::Config,
-        perf_utils::{sample_txs, SampleStats},
         create_nonce::*,
+        perf_utils::{sample_txs, SampleStats},
     },
     log::*,
     rayon::prelude::*,
     solana_client::nonce_utils,
-    solana_core::gen_keys::GenKeys,
-    solana_measure::measure::Measure,
     solana_metrics::{self, datapoint_info},
     solana_sdk::{
         clock::{DEFAULT_MS_PER_SLOT, DEFAULT_S_PER_SLOT, MAX_PROCESSING_AGE},
@@ -20,7 +16,6 @@ use {
         instruction::{AccountMeta, Instruction},
         message::Message,
         native_token::Sol,
-        nonce::State,
         pubkey::Pubkey,
         signature::{Keypair, Signer},
         system_instruction, system_transaction,
@@ -28,11 +23,11 @@ use {
         transaction::Transaction,
     },
     std::{
-        collections::{HashSet, VecDeque},
+        collections::VecDeque,
         process::exit,
         sync::{
             atomic::{AtomicBool, AtomicIsize, AtomicUsize, Ordering},
-            Arc, Mutex, RwLock,
+            Arc, RwLock,
         },
         thread,
         thread::{sleep, Builder, JoinHandle},
@@ -43,10 +38,7 @@ use {
 // The point at which transactions become "too old", in seconds.
 const MAX_TX_QUEUE_AGE: u64 = (MAX_PROCESSING_AGE as f64 * DEFAULT_S_PER_SLOT) as u64;
 
-pub const MAX_SPENDS_PER_TX: u64 = 4;
-
 pub type SharedTransactions = Arc<RwLock<VecDeque<Vec<(Transaction, u64)>>>>;
-
 
 /// Split input vector of keypairs into two sets of chunks of given size
 fn split_into_source_destination(
@@ -257,54 +249,6 @@ where
     }
 }
 
-fn generate_durable_nonce_accounts<T: 'static + BenchTpsClient + Send + Sync>(
-    client: Arc<T>,
-    authority_keypairs: &Vec<Keypair>,
-) -> Result<Vec<Keypair>> {
-    let nonce_rent = client
-        .get_minimum_balance_for_rent_exemption(State::size())
-        .unwrap();
-
-    //let client = client.rpc_client(); // TODO remove later, for now it gives better errors
-    let seed_keypair = &authority_keypairs[0];
-    let count = authority_keypairs.len();
-    let (mut nonce_keypairs, _extra) = generate_keypairs(&seed_keypair, count as u64);
-    nonce_keypairs.truncate(count);
-    let FUND_CHUNK_LEN = 1024; // TODO(klykov) use one everywhere instead
-
-    let to_fund: Vec<(&Keypair, &Keypair)> = authority_keypairs.iter().zip(nonce_keypairs.iter()).map(|pair| {
-        pair
-    }).collect();
- 
-    to_fund.chunks(FUND_CHUNK_LEN).for_each(|chunk| {
-        NonceContainer::with_capacity(chunk.len()).create_accounts(
-                &client,
-                chunk,
-                nonce_rent,
-        );
-    });
-    /*for (authority, nonce) in authority_keypairs.iter().zip(nonce_keypairs.iter()) {
-        // make
-        let instr = system_instruction::create_nonce_account(
-            &authority.pubkey(),
-            &nonce.pubkey(),
-            &authority.pubkey(),
-            nonce_rent,
-        );
-
-        let mut tx = Transaction::new_with_payer(&instr, Some(&authority.pubkey()));
-        // sign
-        let blockhash = get_latest_blockhash(client.as_ref());
-        if tx.try_sign(&[&nonce, authority], blockhash).is_err() {
-            return Err(BenchTpsError::Custom("Signing has failed".to_string()));
-        }
-        // send
-        let result = client.send_transaction(tx)?;
-    }
-    */
-    Ok(nonce_keypairs)
-}
-
 fn withdraw_nonce_account<T: 'static + BenchTpsClient + Send + Sync>(
     client: Arc<T>,
     keypairs: &Vec<Keypair>,
@@ -359,11 +303,11 @@ where
     let (source_nonce_keypair_chunks, mut dest_nonce_keypair_chunks) =
         split_into_source_destination(&nonce_keypairs, tx_count);
 
-    info!("BEFORE");
+    info!("@BEFORE");
     wait_for_commitment(client.clone(), &nonce_keypairs, |lamports: u64| {
         lamports > 0
     });
-    info!("AFTER");
+    info!("@AFTER");
 
     let first_tx_count = loop {
         match client.get_transaction_count() {
@@ -768,222 +712,6 @@ fn do_tx_transfers<T: BenchTpsClient>(
     }
 }
 
-trait FundingTransactions<'a> {
-    fn fund<T: 'static + BenchTpsClient + Send + Sync>(
-        &mut self,
-        client: &Arc<T>,
-        to_fund: &[(&'a Keypair, Vec<(Pubkey, u64)>)],
-        to_lamports: u64,
-    );
-    fn make(&mut self, to_fund: &[(&'a Keypair, Vec<(Pubkey, u64)>)]);
-    fn sign(&mut self, blockhash: Hash);
-    fn send<T: BenchTpsClient>(&self, client: &Arc<T>);
-    fn verify<T: 'static + BenchTpsClient + Send + Sync>(
-        &mut self,
-        client: &Arc<T>,
-        to_lamports: u64,
-    );
-}
-
-impl<'a> FundingTransactions<'a> for Vec<(&'a Keypair, Transaction)> {
-    fn fund<T: 'static + BenchTpsClient + Send + Sync>(
-        &mut self,
-        client: &Arc<T>,
-        to_fund: &[(&'a Keypair, Vec<(Pubkey, u64)>)],
-        to_lamports: u64,
-    ) {
-        self.make(to_fund);
-
-        let mut tries = 0;
-        while !self.is_empty() {
-            info!(
-                "{} {} each to {} accounts in {} txs",
-                if tries == 0 {
-                    "transferring"
-                } else {
-                    " retrying"
-                },
-                to_lamports,
-                self.len() * MAX_SPENDS_PER_TX as usize,
-                self.len(),
-            );
-
-            let blockhash = get_latest_blockhash(client.as_ref());
-
-            // re-sign retained to_fund_txes with updated blockhash
-            self.sign(blockhash);
-            self.send(client);
-
-            // Sleep a few slots to allow transactions to process
-            sleep(Duration::from_secs(1));
-
-            self.verify(client, to_lamports);
-
-            // retry anything that seems to have dropped through cracks
-            //  again since these txs are all or nothing, they're fine to
-            //  retry
-            tries += 1;
-        }
-        info!("transferred");
-    }
-
-    fn make(&mut self, to_fund: &[(&'a Keypair, Vec<(Pubkey, u64)>)]) {
-        let mut make_txs = Measure::start("make_txs");
-        let to_fund_txs: Vec<(&Keypair, Transaction)> = to_fund
-            .par_iter()
-            .map(|(k, t)| {
-                let instructions = system_instruction::transfer_many(&k.pubkey(), t);
-                let message = Message::new(&instructions, Some(&k.pubkey()));
-                (*k, Transaction::new_unsigned(message))
-            })
-            .collect();
-        make_txs.stop();
-        debug!(
-            "make {} unsigned txs: {}us",
-            to_fund_txs.len(),
-            make_txs.as_us()
-        );
-        self.extend(to_fund_txs);
-    }
-
-    fn sign(&mut self, blockhash: Hash) {
-        let mut sign_txs = Measure::start("sign_txs");
-        self.par_iter_mut().for_each(|(k, tx)| {
-            tx.sign(&[*k], blockhash);
-        });
-        sign_txs.stop();
-        debug!("sign {} txs: {}us", self.len(), sign_txs.as_us());
-    }
-
-    fn send<T: BenchTpsClient>(&self, client: &Arc<T>) {
-        let mut send_txs = Measure::start("send_and_clone_txs");
-        let batch: Vec<_> = self.iter().map(|(_keypair, tx)| tx.clone()).collect();
-        client.send_batch(batch).expect("transfer");
-        send_txs.stop();
-        debug!("send {} {}", self.len(), send_txs);
-    }
-
-    fn verify<T: 'static + BenchTpsClient + Send + Sync>(
-        &mut self,
-        client: &Arc<T>,
-        to_lamports: u64,
-    ) {
-        let starting_txs = self.len();
-        let verified_txs = Arc::new(AtomicUsize::new(0));
-        let too_many_failures = Arc::new(AtomicBool::new(false));
-        let loops = if starting_txs < 1000 { 3 } else { 1 };
-        // Only loop multiple times for small (quick) transaction batches
-        let time = Arc::new(Mutex::new(Instant::now()));
-        for _ in 0..loops {
-            let time = time.clone();
-            let failed_verify = Arc::new(AtomicUsize::new(0));
-            let client = client.clone();
-            let verified_txs = &verified_txs;
-            let failed_verify = &failed_verify;
-            let too_many_failures = &too_many_failures;
-            let verified_set: HashSet<Pubkey> = self
-                .par_iter()
-                .filter_map(move |(k, tx)| {
-                    if too_many_failures.load(Ordering::Relaxed) {
-                        return None;
-                    }
-
-                    let verified = if verify_funding_transfer(&client, tx, to_lamports) {
-                        verified_txs.fetch_add(1, Ordering::Relaxed);
-                        Some(k.pubkey())
-                    } else {
-                        failed_verify.fetch_add(1, Ordering::Relaxed);
-                        None
-                    };
-
-                    let verified_txs = verified_txs.load(Ordering::Relaxed);
-                    let failed_verify = failed_verify.load(Ordering::Relaxed);
-                    let remaining_count = starting_txs.saturating_sub(verified_txs + failed_verify);
-                    if failed_verify > 100 && failed_verify > verified_txs {
-                        too_many_failures.store(true, Ordering::Relaxed);
-                        warn!(
-                            "Too many failed transfers... {} remaining, {} verified, {} failures",
-                            remaining_count, verified_txs, failed_verify
-                        );
-                    }
-                    if remaining_count > 0 {
-                        let mut time_l = time.lock().unwrap();
-                        if time_l.elapsed().as_secs() > 2 {
-                            info!(
-                                "Verifying transfers... {} remaining, {} verified, {} failures",
-                                remaining_count, verified_txs, failed_verify
-                            );
-                            *time_l = Instant::now();
-                        }
-                    }
-
-                    verified
-                })
-                .collect();
-
-            self.retain(|(k, _)| !verified_set.contains(&k.pubkey()));
-            if self.is_empty() {
-                break;
-            }
-            info!("Looping verifications");
-
-            let verified_txs = verified_txs.load(Ordering::Relaxed);
-            let failed_verify = failed_verify.load(Ordering::Relaxed);
-            let remaining_count = starting_txs.saturating_sub(verified_txs + failed_verify);
-            info!(
-                "Verifying transfers... {} remaining, {} verified, {} failures",
-                remaining_count, verified_txs, failed_verify
-            );
-            sleep(Duration::from_millis(100));
-        }
-    }
-}
-
-/// fund the dests keys by spending all of the source keys into MAX_SPENDS_PER_TX
-/// on every iteration.  This allows us to replay the transfers because the source is either empty,
-/// or full
-pub fn fund_keys<T: 'static + BenchTpsClient + Send + Sync>(
-    client: Arc<T>,
-    source: &Keypair,
-    dests: &[Keypair],
-    total: u64,
-    max_fee: u64,
-    lamports_per_account: u64,
-) {
-    let mut funded: Vec<&Keypair> = vec![source];
-    let mut funded_funds = total;
-    let mut not_funded: Vec<&Keypair> = dests.iter().collect();
-    while !not_funded.is_empty() {
-        // Build to fund list and prepare funding sources for next iteration
-        let mut new_funded: Vec<&Keypair> = vec![];
-        let mut to_fund: Vec<(&Keypair, Vec<(Pubkey, u64)>)> = vec![];
-        let to_lamports = (funded_funds - lamports_per_account - max_fee) / MAX_SPENDS_PER_TX;
-        for f in funded {
-            let start = not_funded.len() - MAX_SPENDS_PER_TX as usize;
-            let dests: Vec<_> = not_funded.drain(start..).collect();
-            let spends: Vec<_> = dests.iter().map(|k| (k.pubkey(), to_lamports)).collect();
-            to_fund.push((f, spends));
-            new_funded.extend(dests.into_iter());
-        }
-
-        // try to transfer a "few" at a time with recent blockhash
-        //  assume 4MB network buffers, and 512 byte packets
-        const FUND_CHUNK_LEN: usize = 4 * 1024 * 1024 / 512;
-
-        to_fund.chunks(FUND_CHUNK_LEN).for_each(|chunk| {
-            Vec::<(&Keypair, Transaction)>::with_capacity(chunk.len()).fund(
-                &client,
-                chunk,
-                to_lamports,
-            );
-        });
-
-        info!("funded: {} left: {}", new_funded.len(), not_funded.len());
-        funded = new_funded;
-        funded_funds = to_lamports;
-    }
-}
-
 fn compute_and_report_stats(
     maxes: &Arc<RwLock<Vec<(String, SampleStats)>>>,
     sample_period: u64,
@@ -1049,22 +777,6 @@ fn compute_and_report_stats(
         "\tAverage TPS: {}",
         max_tx_count as f32 / duration_as_s(tx_send_elapsed)
     );
-}
-
-pub fn generate_keypairs(seed_keypair: &Keypair, count: u64) -> (Vec<Keypair>, u64) {
-    let mut seed = [0u8; 32];
-    seed.copy_from_slice(&seed_keypair.to_bytes()[..32]);
-    let mut rnd = GenKeys::new(seed);
-
-    let mut total_keys = 0;
-    let mut extra = 0; // This variable tracks the number of keypairs needing extra transaction fees funded
-    let mut delta = 1;
-    while total_keys < count {
-        extra += delta;
-        delta *= MAX_SPENDS_PER_TX;
-        total_keys += delta;
-    }
-    (rnd.gen_n_keypairs(total_keys), extra)
 }
 
 pub fn generate_and_fund_keypairs<T: 'static + BenchTpsClient + Send + Sync>(
