+++ bench-tps/src/lib.rs
@@ -2,6 +2,6 @@
 pub mod bench;
 pub mod bench_tps_client;
 pub mod cli;
-mod create_nonce;
+pub mod create_nonce;
 pub mod keypairs;
 mod perf_utils;
