+++ bench-tps/src/main.rs
@@ -3,8 +3,9 @@ use {
     clap::value_t,
     log::*,
     solana_bench_tps::{
-        bench::{do_bench_tps, generate_keypairs},
+        bench::do_bench_tps,
         cli::{self, ExternalClientType},
+        create_nonce::generate_keypairs,
         keypairs::get_keypairs,
     },
     solana_client::{
